mining_gems.lua
Reduce 6s delay to 3s for fail safe continuation.
Many tales ago we were using 6s because the game was crazy lagging.
This Tale 11, we have a very stable 0.89 - 0.9 ratio and very lag free.
We can reduce this fail safe by 50%; I've tested this for several hours.

Created variable maxWaitContinue to denote this; It can be manually tweaked if needed.

Reduced clickDelay from 150 to 100. I can easily run 60ms without any issues this tale. But we'll keep this 100 to be safe.